[MPDF-84] display pdf generated file(s) location
git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@1816336 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-pdf-plugin/src/main/java/org/apache/maven/plugins/pdf/PdfMojo.java
@@ -449,10 +449,7 @@ public class PdfMojo
     private void copyGeneratedPdf()
         throws MojoExecutionException, IOException
     {
-        if ( outputDirectory.getCanonicalPath().equals( workingDirectory.getCanonicalPath() ) )
-        {
-            return;
-        }
+        boolean requireCopy = !outputDirectory.getCanonicalPath().equals( workingDirectory.getCanonicalPath() );
 
         String outputName = getDocumentModel( getDefaultLocale() ).getOutputName().trim();
         if ( !outputName.endsWith( ".pdf" ) )
@@ -472,8 +469,13 @@ public class PdfMojo
 
             File generatedPdfDest = new File( getLocaleDirectory( outputDirectory, locale ), outputName );
 
-            FileUtils.copyFile( generatedPdfSource, generatedPdfDest );
-            generatedPdfSource.delete();
+            if ( requireCopy )
+            {
+                FileUtils.copyFile( generatedPdfSource, generatedPdfDest );
+                generatedPdfSource.delete();
+            }
+
+            getLog().info( "pdf generated: " + generatedPdfDest );
         }
     }
